pyproject.toml
[project]
name = "ai-interviewer-mvp"
version = "0.1.0"
description = "Desktop MVP for conducting AI-assisted Google Meet interviews."
authors = [{ name = "AI" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6",
    "sounddevice>=0.4",
    "soundfile>=0.12",
    "pyttsx3>=2.90",
    "playwright>=1.42",
    "PyPDF2>=3.0",
    "python-docx>=1.0",
    "faster-whisper>=0.10",
    "openai>=1.30",
    "anthropic>=0.20",
    "google-generativeai>=0.5",
]

[project.scripts]
ai-interviewer = "app.main:main"

[tool.playwright]
automation-install = true
